+++ src/js/handler/handleWinningNumberInput.js
@@ -1,34 +1,8 @@
 import { $, $$ } from '../utils/querySelector.js';
-import { isDuplicate, isValidRange } from '../utils/validator.js';
 import { ERR_MESSAGE, VALUE } from '../utils/constant.js';
+import { isDuplicate, isValidRange } from '../utils/validator.js';
 import { openModal, renderModal } from '../view/viewModalPage.js';
-
-const getProfit = (winningRank) => {
-  const profits = {
-    [VALUE.WINNING_RANK.FIRST]: VALUE.WINNING_PRICE.FIRST,
-    [VALUE.WINNING_RANK.SECOND]: VALUE.WINNING_PRICE.SECOND,
-    [VALUE.WINNING_RANK.THIRD]: VALUE.WINNING_PRICE.THIRD,
-    [VALUE.WINNING_RANK.FOURTH]: VALUE.WINNING_PRICE.FOURTH,
-    [VALUE.WINNING_RANK.FIFTH]: VALUE.WINNING_PRICE.FIFTH,
-    [VALUE.WINNING_RANK.NONE]: VALUE.WINNING_PRICE.NONE,
-  };
-
-  return profits[winningRank];
-};
-
-const getRank = (winningCount) => {
-  const rank = {
-    [VALUE.HIT_COUNT.SIX]: VALUE.WINNING_RANK.FIRST,
-    [VALUE.HIT_COUNT.FIVE]: VALUE.WINNING_RANK.THIRD,
-    [VALUE.HIT_COUNT.FOUR]: VALUE.WINNING_RANK.FOURTH,
-    [VALUE.HIT_COUNT.THREE]: VALUE.WINNING_RANK.FIFTH,
-    [VALUE.HIT_COUNT.TWO]: VALUE.WINNING_RANK.NONE,
-    [VALUE.HIT_COUNT.ONE]: VALUE.WINNING_RANK.NONE,
-    [VALUE.HIT_COUNT.NONE]: VALUE.WINNING_RANK.NONE,
-  };
-
-  return rank[winningCount];
-};
+import { getTicketResult } from '../components/getTicketResult.js';
 
 const getRankCountMap = (lotto) => {
   const rankCountMap = new Map([
@@ -55,20 +29,17 @@ const getTotalYield = (lotto) => {
   return Number(((totalProfit / lotto.purchasePrice) * 100).toFixed(2));
 };
 
-const getTicketResult = (ticket, winningNumbers, bonusNumber) => {
-  const bonusCount = ticket.numbers.includes(bonusNumber);
-  const winnigCount =
-    VALUE.LOTTO.TICKET_LENGH * 2 -
-    new Set([...ticket.numbers, ...winningNumbers]).size;
+const isValidNumberInput = (inputNumbers) => {
+  if (!isValidRange(inputNumbers)) {
+    alert(ERR_MESSAGE.WINNING_NUMBER.OUT_OF_RANGE);
+    return false;
+  }
 
-  const winningRank =
-    bonusCount && winnigCount === VALUE.HIT_COUNT.FIVE
-      ? VALUE.WINNING_RANK.SECOND
-      : getRank(winnigCount);
-  const profit = getProfit(ticket.winningRank);
-
-  ticket.setWinningRank(winningRank);
-  ticket.setProfit(profit);
+  if (isDuplicate(inputNumbers)) {
+    alert(ERR_MESSAGE.WINNING_NUMBER.DUPLICATE);
+    return false;
+  }
+  return true;
 };
 
 export const handleWinningNumberInput = (lotto) => {
@@ -76,22 +47,16 @@ export const handleWinningNumberInput = (lotto) => {
     Number(winningNumber.value),
   );
   const bonusNumber = Number($('.bonus-number').value);
+  const inputNumbers = [...winningNumbers, bonusNumber];
 
-  if (!isValidRange([...winningNumbers, bonusNumber])) {
-    return alert(ERR_MESSAGE.WINNING_NUMBER.OUT_OF_RANGE);
-  }
-
-  if (isDuplicate([...winningNumbers, bonusNumber])) {
-    return alert(ERR_MESSAGE.WINNING_NUMBER.DUPLICATE);
+  if (!isValidNumberInput(inputNumbers)) {
+    return;
   }
 
   lotto.tickets.forEach((ticket) => {
     getTicketResult(ticket, winningNumbers, bonusNumber);
   });
 
-  const ranckCountMap = getRankCountMap(lotto);
-  const totalYield = getTotalYield(lotto);
-
-  renderModal(ranckCountMap, totalYield);
+  renderModal(getRankCountMap(lotto), getTotalYield(lotto));
   openModal();
 };
